Rename appcontainer.rs to appcontainer_runner.rs
Rename module to follow the _runner.rs naming convention for files
containing runner classes. Update module declaration in lib.rs and
import in main.rs accordingly.
